move ScriptModuleDescrializer template to a header file
